R Move `KrakenImageViewController` to its own file
